new header structure
should lead to fewer total rebuilds
