--- include/Libs.hpp
@@ -1,26 +1,0 @@
-/**
- * \file Libs.hpp
- * \brief File with libraries and dependencies
- * \author GenaBitu
- **/
-
-#ifndef LIBS_HPP
-#define LIBS_HPP
-
-#include <iostream>
-#include <fstream>
-#include <vector>
-#include <thread>
-#include <memory>
-#include <iterator>
-#include <GL/glew.h>
-#include <GLFW/glfw3.h>
-#define GLM_FORCE_RADIANS
-#include <glm/glm.hpp>
-#include <glm/gtc/matrix_transform.hpp>
-#include <glm/gtc/quaternion.hpp>
-#include <glm/gtc/type_ptr.hpp>
-#include <ft2build.h>
-#include FT_FREETYPE_H
-
-#endif // LIBS_HPP
