+++ include/TextureBMP.hpp
@@ -6,7 +6,7 @@
 
 #ifndef TEXTUREBMP_HPP
 #define TEXTUREBMP_HPP
-#include "Libs.hpp"
+#include "Global.hpp"
 #include "TextureBridge.hpp"
 
 #pragma pack(push)
