+++ src/Text.cpp
@@ -1,4 +1,6 @@
-#include "../include/Main.hpp"
+#include "../include/Text.hpp"
+
+#include "../include/LineBreak.hpp"
 using namespace std;
 using namespace glm;
