+++ src/Texture.cpp
@@ -1,4 +1,4 @@
-#include "../include/Main.hpp"
+#include "../include/Texture.hpp"
 using namespace std;
 using namespace glm;
